LayoutTests/platform/mac-mojave/fast/text/international/system-language/navigator-language/navigator-language-es-419-expected.txt

FAIL navigator.language should be es-419. Was es-XL.
PASS successfullyParsed is true
Some tests failed.

TEST COMPLETE
